// src/app/common/model/enums/form.enums.ts
export enum ControlTypeEnum {
	FormHidden,
	FormControl,
	FormArray,
	FormGroup,
}

export enum FormControlTypeEnum {
	INPUT,
	AUTOCOMPLETE,
	CHECKBOX,
	DATEPICKER,
	IMAGE,
	MULTISELECT,
	RADIO,
	SELECT,
	SLIDER,
	TEXTAREA,
	TOGGLE,
}

export enum InputTypeEnum {
	text = 'text',
	color = 'color',
	date = 'date',
	datetimeLocal = 'datetime-local',
	email = 'email',
	month = 'month',
	number = 'number',
	password = 'password',
	search = 'search',
	tel = 'tel',
	time = 'time',
	url = 'url',
	week = 'week',
}

export enum InputFieldSuffix {
	day = 'dzien',
	cm = 'cm',
	mm = 'mm',
	price = '.00 zł',
}